changed int type to int64 for GCD, added LCM, updated readme

+++ gcd.go
@@ -1,13 +1,13 @@
 package algorithms
 
-func max(a, b int) int {
+func max(a, b int64) int64 {
 	if a >= b {
 		return a
 	}
 	return b
 }
 
-func min(a, b int) int {
+func min(a, b int64) int64 {
 	if a <= b {
 		return a
 	}
@@ -15,7 +15,7 @@ func min(a, b int) int {
 }
 
 // GCD returns gcd of a and b.
-func GCD(a, b int) int {
+func GCD(a, b int64) int64 {
 	a1 := max(a, b)
 	b1 := min(a, b)
 	for b1 != 0 {
